+++ app/src/main/AndroidManifest.xml
@@ -1,12 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.julia.bachelor">
-
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
     <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
     <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
-
     <application
         android:allowBackup="true"
         android:icon="@drawable/logo"
@@ -22,14 +20,13 @@
             android:label="@string/title_activity_main">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
+                <action android:name="android.intent.action.VIEW" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
         <activity android:name=".SalgItem" />
         <activity android:name=".FullOversikt" />
         <activity android:name=".SalgAnnet" />
-        <activity android:name=".AddBeholdning" />
         <activity android:name=".FakturaSalg" />
         <activity android:name=".BmSalg" />
         <activity android:name=".HjemmeSalg" />
@@ -46,5 +43,4 @@
                 android:resource="@xml/provider_paths"/>
         </provider>
     </application>
-
 </manifest>
