+++ app/src/main/res/layout/fragment_skal_bli_add_beholdning.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#f3f3f3"
-    tools:context=".SkalBliAddBeholdning">
+    tools:context=".AddBeholdning">
 
     <ScrollView
         android:layout_width="match_parent"
